# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Fallstudie-Gruppe-3",
    "description": "API Dokumentation zur Anbindung von Frontends für das Kinoticketreservierungssystem."
  },
  "servers": [
    {
      "url": "https://fallstudie-gruppe-3.herokuapp.com/"
    },
    {
      "url": "http://fallstudie-gruppe-3.herokuapp.com/"
    }
  ],
...
  "paths": {
    "/filme": {
      "get": {
        "tags": [
          "Film"
        ],
        "description": "Alle Filme abrufen.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Film"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Film"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titel": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "description": "Film anlegen.",
        "tags": [
          "Film"
        ]
      }
    },
    "/filme/{titel}": {
      "delete": {
        "tags": [
          "Film"
        ],
        "description": "Film löschen.",
        "parameters": [
          {
            "name": "titel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/saele": {
      "post": {
        "tags": [
          "Saal"
        ],
        "description": "Saal anlegen.",
        "parameters": [
          {
            "name": "newHall",
            "in": "body",
            "description": "Anzulegender Saal.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SaalAnlegen"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/saele/{nummer}": {
      "delete": {
        "tags": [
          "Saal"
        ],
        "description": "Saal löschen.",
        "parameters": [
          {
            "name": "nummer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/sitzplaetze/reservieren": {
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entries": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "description": "Sitzplatz reservieren.",
        "tags": [
          "Sitzplatz"
        ]
      }
    },
    "/sitzplaetze/checkout": {
      "put": {
        "tags": [
          "Sitzplatz"
        ],
        "description": "Sitzplatz buchen.",
        "parameters": [
          {
            "name": "seats",
            "in": "body",
            "description": "Zu buchende Sitze.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SitzeBuchen"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/sitzplaetze/freigeben": {
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entries": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "description": "Sitzplatz freigeben.",
        "tags": [
          "Sitzplatz"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Film": {
        "type": "object",
        "properties": {
          "titel": {
            "type": "string",
            "example": "Inception"
          },
          "bild": {
            "type": "string",
            "example": "https://imdb-api.com/images/original/MV5BMjAxMzY3NjcxNF5BMl5BanBnXkFtZTcwNTI5OTM0Mw@@._V1_Ratio0.6762_AL_.jpg"
          },
          "dauer": {
            "type": "number",
            "example": 148
          },
          "beschreibung": {
            "type": "string",
            "example": "Beim Versuch in das Unterbewusstsein des Industriellen Saito einzudringen, stoßen Cobb und seine Mitstreiter auf unerwartete Schwierigkeiten. Saito hatte Verteidigungsstrategien gegen den Gedankendiebstahl trainiert und baut nun einen Abwehrriegel auf, doch erst der Verrat eines Kollegen lässt den Versuch scheitern, dem Schlafenden Geheimnisse zu entlocken. Da seine Auftraggeber unbedingten Erfolg gefordert hatten, muss Cobb, der auf Grund einer drohenden Verurteilung nicht mehr in die USA zurück kann, fliehen, doch Saito kommt ihm zuvor. Allerdings nicht um ihn zu bestrafen, sondern um ihm einen Deal vorzuschlagen."
          },
          "trailer": {
            "type": "string",
            "example": "https://www.youtube.com/embed/dHTyZ9Bmp0c"
          },
          "vorstellungen": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "saal": {
                  "type": "number",
                  "example": 1
                },
                "startzeit": {
                  "type": "string",
                  "example": "2021-12-19T16:00"
                },
                "sitzplaetze": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "reihe": {
                        "type": "string",
                        "example": "A"
                      },
                      "nummer": {
                        "type": "number",
                        "example": 1
                      },
                      "sitzart": {
                        "type": "string",
                        "example": "einzel"
                      },
                      "reserviert": {
                        "type": "boolean",
                        "example": false
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "Film"
        }
      },
...
      "SaalAnlegen": {
        "type": "object",
        "properties": {
          "nummer": {
            "type": "number",
            "example": 1
          },
          "sitzplaetze": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "reihe": {
                  "type": "string",
                  "example": "A"
                },
                "nummer": {
                  "type": "number",
                  "example": 1
                },
                "sitzart": {
                  "type": "string",
                  "example": "einzel"
                },
                "reserviert": {
                  "type": "boolean",
                  "example": false
                }
              },
              "required": [
                "reihe",
                "nummer",
                "sitzart",
                "reserviert"
              ]
            }
          }
        },
        "required": [
          "nummer",
          "sitzplaetze"
        ],
        "xml": {
          "name": "SaalAnlegen"
        }
      },
...
      "SitzeBuchen": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "titel": {
              "type": "string",
              "example": "Spider-Man: No Way Home"
            },
            "saal": {
              "type": "number",
              "example": 1
            },
            "startzeit": {
              "type": "string",
              "example": "2022-01-05T12:30"
            },
            "reihe": {
              "type": "string",
              "example": "A"
            },
            "nummer": {
              "type": "number",
              "example": 15
            },
            "wert": {
              "type": "string",
              "example": "auth0|userId"
            }
          },
          "required": [
            "titel",
            "saal",
            "startzeit",
            "reihe",
            "nummer",
            "wert"
          ]
        },
        "xml": {
          "name": "SitzeBuchen"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}